requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
python-dotenv>=1.0
gunicorn>=20.1
requests>=2.30
pywa
whitenoise